fix(android): 修复 Kotlin 编译错误
- PrivilegedExecutionModule: Float progress 转 Double 供 putDouble
- VoiceInteractionService: ERROR_NO_MATCH 常量需用 RecognitionService.Callback 引用

+++ mobile/native-android/kotlin/com/monkeycode/privileged/PrivilegedExecutionModule.kt
@@ -376,7 +376,7 @@ class PrivilegedExecutionModule(reactContext: ReactApplicationContext) :
             try {
                 alpineEnvironment.install { progress ->
                     sendEvent("alpineInstallProgress", Arguments.createMap().apply {
-                        putDouble("progress", progress)
+                        putDouble("progress", progress.toDouble())
                     })
                 }
                 val result = Arguments.createMap()
